refactor: split user settings routes
